+++ Assets/Scripts/VoxelModel.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 32ed41ede89a94dbabea910e1b2aace1
+folderAsset: yes
+timeCreated: 1513407316
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
